re-work proprietary blob extraction
AOSP has re-worked the way it distributes proprietary drivers.
This patch looks for a file under vendor/columbia/build/PRODUCT/
named "PRODUCT_binaries.tar.gz" and extracts it if needed.

For example, the Asus Nexus 7 has a product name of "grouper"
so the script would look for:
    vendor/columbia/build/grouper/grouper_binaries.tar.gz
and extract it if the directory
    "vendor/asus" doesn't already exist.

Change-Id: I688ef2a33f4c70b33e87462a5b6e29b60f745660
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>