cells: custom init.rc files for host and cell
Tell the Android build system that we provide a custom init.rc file.

The Cells init.rc has been trimmed down to support a minimal
command-line interface for the host. The init.cell.rc file contains an
init.rc file suitable for starting a guest Cell. It is the
responsibility of CellD to manage cell startup and use of the proper
init file through a custom hook into the Android init process.

Change-Id: Id42d895d4192e5609e7a66380803d232a447ee9b
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>